Reverted unnecessary fix :rage4:
Apparently the bot has issues processing too many messages in a
short ammount of time and this was causing most of the crashes
in my deploy. The instance I have is not really "powerful" and
I guess this is what causing the issues and I'll just revert the
changes I made before.