Fix: improved the error log and added guard
The error object being sent naively to "printError" would not print the
stack trace. I decided it was better for that method to print a simple
error message but not the object, instead using "console.error". On the
other hand, I realized printing the full context object was not
necessary, instead it's better to just print the chat and update
properties from it. Lastly, I noticed that the error being triggered was
due to the bot receiving the "edit message" event and handling it badly
on the middleware function (the message object was undefined), this
should fix that error too.